refactor: 디렉토리 구조화, 컴포넌트 네이밍 리팩토링
