+++ client/src/pages/HomePage.vue
